Don't quote subquery columns to prevent column mismatches on multi-column references
For simple collection queries where the junction table is not interrogated,
an `IN` expression is used in the `WHERE` clause to filter the right table
to records that are in the junction table. In cases where the right table
has more than one reference columns, the subquery puts the columns in parenthesis
which causes them to be treated as a single compound column rather than a tuple.
This results in an error due to a mismatch in the number of columns on either
side of the expression.